Switch npm test to using mocha directly (#54)
Remove gulp-mocha-phantomjs dependency and run tests without a
browser context. Loading test.html in a browser still succeeds.

Remove gulp test task.

Update to latest chai version now phantomjs is not involved.

Introduce setup file for tests that requires the needed components
and populates the global object. This can be folded back into the
test file later and a browser importable version generated as a step
in the build process. For now it needs some lint overrides.

Only handle requests in the mock XMLHttpRequest implementation which
match the endpoint url.